backend/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
google-generativeai==0.3.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.3
gunicorn==21.2.0
pytest==7.4.3
lxml==4.9.3
PyPDF2==3.0.1
